update sidemenu tests
